feat: Add Freight & Logistics module and Native Product Translation

+++ migrations/0031_document_number_counters.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS document_number_counters (
   company_id integer NOT NULL,
   prefix text NOT NULL,
-  last_number integer NOT NULL DEFAULT 0,
+  last_number bigint NOT NULL DEFAULT 0,
   updated_at timestamp DEFAULT now() NOT NULL,
   PRIMARY KEY (company_id, prefix)
 );
